backend/requirements.txt

# Django Core
Django
asgiref

# Django REST Framework
djangorestframework
djangorestframework-simplejwt

# CORS
django-cors-headers

# PostgreSQL
psycopg2-binary

# Variables de entorno
python-dotenv

# Utilidades
pytz
sqlparse
qrcode
celery[redis,django]

ruff